tear down of sanity check test
